+++ comics/comics/pennyarcade.py
@@ -5,7 +5,7 @@ from comics.core.comic_data import ComicDataBase
 class ComicData(ComicDataBase):
     name = 'Penny Arcade'
     language = 'en'
-    url = 'http://www.penny-arcade.com/'
+    url = 'http://penny-arcade.com/'
     start_date = '1998-11-18'
     rights = 'Mike Krahulik & Jerry Holkins'
 
@@ -15,8 +15,11 @@ class Crawler(CrawlerBase):
     schedule = 'Mo,We,Fr'
     time_zone = 'US/Pacific'
 
+    # Without User-Agent set, the server returns 403 Forbidden
+    headers = {'User-Agent': 'Mozilla/4.0'}
+
     def crawl(self, pub_date):
-        page_url = 'http://www.penny-arcade.com/comic/%s/' % (
+        page_url = 'http://penny-arcade.com/comic/%s' % (
             pub_date.strftime('%Y/%m/%d'),)
         page = self.parse_page(page_url)
         title = page.alt('#comicFrame img')